pyproject.toml
[tool.poetry]
name = "photoshare"
version = "0.1.0"
description = ""
authors = ["N0yhz <vutunhzyonh@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = {extras = ["email"], version = "^2.10.4"}
fastapi = "^0.115.6"
slowapi = "^0.1.9"
fastapi-mail = "^1.4.2"
sqlalchemy = "^2.0.36"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
alembic = "^1.14.0"
qrcode = "^8.0"
pillow = "^11.0.0"
requests = "^2.32.3"
greenlet = "^3.1.1"
asyncpg = "^0.30.0"
cloudinary = "^1.41.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.20"
redis = "^5.2.1"
bcrypt = "^3.1.7"
passlib = "^1.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"